chore: validate addons parameters before deployment (#4361)
This PR validates the parameters in addon templates, as well as the customized parameters, before the CFN deployment. 

With the wrong addons parameter config, the CFN deployment fails nevertheless. With this PR, Copilot will error out earlier so that users don't have to wait until the deployment to know.

Note that this validation happens for both the workload addon, and the environment addon.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.