fix: Better error message for alias validation against imported certificate (#4250)
Currently when customer import a certificate and specify alias in their service manifest, Copilot validates the aliases against that imported certificate. For example, if the certificate allows `images.example.com`, then an alias of `test.images.example.com` is not valid.

Copilot error out saying something like `"test.images.example.com" is not a valid domain against <certificate ARN>`.  But we don’t really say what’s in the certificate.  

With this PR we log out the  details  of the imported `certificate ARN` and all the allowed domains in the Certificate.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.