fix(custom-domain-action): ignore A-record deletion if not found (#4409)
Fixes #4401

If the record fails to insert, then the custom resource fails. On failure, it tries to delete the record that was *not* inserted. Hence the Delete workflow fails resulting in the custom resource to fail to delete.

This change ensures that we ignore the error if the A-record that we're trying to delete doesn't exist in the first place.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.