chore(templates): write ingress as an RDWS workload addon for environment-level aurora (#4393)
For an RDWS to access an environment-level aurora serverless cluster, we need to attach to `ServiceSecurityGroup` with an ingress for access to the cluster.

This ingress is created as a hook resource in the RDWS's workload-level addon.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.