refactor: deploy and template pkg to modify portmappings config (#4357)
This PR focuses on converting exposedPorts into portMappings in order to set right set of ports in the portmappings for primary as well as sidecar containers. 

This PR is a part of https://github.com/aws/copilot-cli/pull/4192



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.