feat: add `--add-ingress-from` flag to `storage init` (#4468)
The flag will produce the `AccessPolicy` or  `SecurityGroupIngress` as a workload-level addon for the workload to access an env-level storage.

This PR includes the `Validate` and `Ask` part of the flag.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.