refactor(deploy): use test double in tests instead of init a stack (#4472)
Today, we're creating a real `stack` object in our `cli/deploy` unit tests. When modifying the constructors in the `stack` pkg that results in all the tests in `cli/deploy` to fail because some dependency goes missing.

This change allows the unit tests in `cli/deploy` to pass in a stub CFN stack instead of initializing a real one.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.