fix(template): disassociate abr when delete (#4565)
This PR fixes a bug that when users remove ABR from their manifests and run `svc deploy`, the alarms will not be disassociated from the services. Also, `svc deploy` to remove ABR will fail if the alarm is currently in `ALARM` state.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.