chore: assign priority to multiple listener rules (#4568)
This PR fixes assigning the priority to multiple listener rules in the service. But it does not solve the problem of assigning priorities to two different services being deployed at once. That case has to be taken care in the future. 


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.