feat: add env file support for all sidecar containers (#4567)
- chore: add Env File field to manifest and add mft validation
- chore: change how env files are handled in parameters
- chore: change how manifest package handles env files
- chore: update templates to include required policies and Env File keys
- chore: add unit tests for cli/deploy package
- chore: fix cli deploy fields
- test: fix stack integ tests


This PR adds support for env files in all customer-defined sidecars, as well as the managed 
Logging sidecar pattern. 

It does this by creating a new parameter for each sidecar, ${sidecarName}EnvFileARN, and 
the corresponding condition. It uses that condition to modify the execution role policy 
so that the ECS service has access to all env files used in the task. 

The feature uploads each env file only once, allowing customers to specify the same env file
across multiple containers and not waste any space in S3 or bandwidth uploading duplicates.


Addresses #4447.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.