chore(progtrack): add rollback alarm names & states (#4563)
These changes add a table of the rollback alarms associated with the deploying service and their states.

This is what it looks like mid-deployment:
![Screenshot 2023-02-28 at 10 38 46 AM](https://user-images.githubusercontent.com/60631893/221990439-f6ccf3b5-15bf-437d-a95d-29cd14fd8692.png)

This is what it looks like after deployment:
![Screenshot 2023-02-28 at 1 58 19 PM](https://user-images.githubusercontent.com/60631893/221990853-5d39ed62-4de7-4715-8508-3a023330bd74.png)

With multiple alarms:
![Screenshot 2023-02-28 at 3 13 00 PM](https://user-images.githubusercontent.com/60631893/222004347-055cc8c8-973c-46e4-8a0c-1e0af85606c3.png)



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.