chore: implements YAML diff writer for `map` and scalar `seq` (#4634)
That writer works for `map` node, `scalar` node and  `seq` nodes whose content are not `map`.

The writer is on parity with the differ in `diff.go`. The to do for both are the same:
1. Handle "(x unchanged items)"
2. Handle `seq` with `map` content


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.