chore: add `--diff` flag in svc/job/env deploy (#4657)
Looks like a large PR but the logic is super similar across env/svc/job deploy. Most of the additions are because of unit tests! (and the unit tests are similar across env/svc/job deploy as well)

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.