chore(describe): fix `svc show` to list only rollback alarms (#4667)
Previously, we erroneously checked for rollback names to be not nil, then fetched the descriptions of the (non-existent) names. The `DescribeAlarms` call fetches all alarms in an account if no alarm names are passed in as a filter. So if no rollback alarms were present, all non-rollback alarm descriptions would be displayed!

Instead, the condition should check for a non-zero slice length because the empty slice is not nil.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.