chore: perform `docker login` only once for all the images (#4697)
In the previous PR #4653 (which is reverted) is printing the below messages even when we run `copilot svc override` as shown below.
![image](https://user-images.githubusercontent.com/71282729/228330219-49fc208f-7d1b-47ce-8dca-477a95e583fa.png)

This PR will fix that issue by removing the  `docker login` process  in `workloadDeployer` constructor and  instead perform `docker login` just before the Build and Push of the images occurs.  In this way we ensure that we print `Login Succeeded` only if we build and push any images.

Summary of the changes.

1. Refactor: move `dockerCmdClient` to Repository struct.
2. Perform `docker Login` just before `BuildAndPush` of images occurs for all `workloads` and `task`.
3. Remove [r.URI()](https://github.com/aws/copilot-cli/blob/mainline/internal/pkg/cli/task_run.go#L730) in `task_run.go` and instead use the `uri` from `repository.Login()`.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.