chore: upload static site files to cache (#4722)
This PR makes the static site files upload to the app bucket with the key `/local-assets/<file hash>`. It also uploads a mapping JSON file mapping each hashed file to a destination location in the static site bucket. It then passes that location of the mapping file to the template package. In a future PR, the mapping file will passed to the template package to create a step machine+lambda that will copy the cached files over to a new bucket for the static site.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.