fix: do record updates in parallel (#4772)
This PR implements the "small" changes in #4711, running the Custom Domain record updates in parallel. It does not implement the large changes, which would imply modifying the cert rather than deleting and recreating it. 

Fixes #4711 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.