fix: change DNS alias in cloudfront test to match cert (#4783)
This PR fixes a recurring issue with the Cloudfront e2e test. The Certificate created in the test does not match the dns alias of the frontend service, so the CF deployment fails. 


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.