fix: don't add the ABR struct to the service if the manifest doesn't call for it (#4782)
This PR fixes #4778. It seems ABR is not available in GovCloud, so all service deployments with 1.27.0 into GovCloud regions fail. 

This PR changes the alarm injection logic to only add the ABR config if it's specified in the manifest, and exercises it one more time
in an integ test. 




By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.