chore(static-site): hash uploaded content for asset mapping file (#4785)
This change makes it so that the **asset mapping file name** is now a **hash of all the uploaded content**. In the template package (see https://github.com/aws/copilot-cli/pull/4762), the path to that file is passed to the custom resource that triggers the copy state machine. By making the file name based on a hash of the content, the state machine will only be triggered if any of the uploaded content changes.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.