fix: use the more portable CDK executable for cdk overrides (#4808)
The old one points to a script with shebang line, which is ignored by Windows. As a result, when I used cdk overrides on Windows, it ends up not being able to interprete the cdk.js file

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.