chore: do not create ecr for static site (#4906)
Fixing this by adding one boolean parameter to `AddWkldToApp` to indicate if we need to create ECR or not. Sounds like a  non-scalable solution but I don't see a good reason why we want to continue adding more resources per service in stackset instance, as it is not compatible if we want to support `svc deploy` from manifest in the future. Also this way (how we manage ECR) strengthen the coupling between app and services.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.