chore: add methods to conditionally remove stackset instance from app (#5060)
PR 1 of 2 replacing #4992.

This adds support methods and a whole boatload of mocks and unit tests to the stackset client and Cloudformation packages. These are necessary to conditionally remove the stackset instance, or redeploy the stackset instances & app to remove an account from dns delegation and IAM roles.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.