chore: infer package manager when overriding with CDK (#5076)
Will have to rebase on #5075

Resolve parts of #5050 on using yarn instead of npm for cdk override, with a small change from the [the first part of the comment I gave](https://github.com/aws/copilot-cli/issues/5050#issuecomment-1625851258).


I decided not include the drafted proposal of "reading package manager selection from env var"; instead, this PR just defaults to `npm` if after all the checks, there is still a tie. This is because I am not clear on whether it is a common design practice to use env vars to decide which package manager to use, and I expect that the current solution (check install + check project)  is already enough to resolve most of the case.


We can always add the enhancement of "reading package manager selection from env var" afterwards, if it's deemed needed.


This does not handle the second ask from the user, i.e. not assuming the location of `node_modules/`.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.