chore: enable ctrl-c only after changeset is created and executed (#5112)
Even though this PR looks big, half of the changes are just unit tests.

Better solution than #5097 

**Old Approach:**
I closed PR 5097. In the previous PR #5097, there is a drawback with the logic when users initiated a ctrl-c interruption during the process of `proposing infrastructure changes`. The issue was that the interruption signal (SIGINT) was being captured even before the changeset is created and executed. As a result, when the user pressed ctrl-c during `proposing infrastructure changes`, we display  a message like `Received interrupt for ctrl-c`; However, despite the interruption, the code continued to create and execute the changeset, leading to the stack creation process continuing even after the user intended to stop it.


**New Approach:**
To address the  issue, this PR had a better solution for handling ctrl-c interruptions during the process of proposing infrastructure changes. Listening to the SIGINT signal is performed only after the changeset is created and executed. The necessary changes to achieve this have been made by relocating the complete logic for waiting and handling the signal to the `deploy/cloudformation` package.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.