chore: fix flaky unittests because of async execution (#5135)
In the mean time I will work on the fix for the test cases. I will revert these changes
I do not think there is any easy way to test the behavior of `ctrl-c`. Because signal can be received at any time and the Stack creation or update can be successful before receiving the signal and leading to missing calls to some of mocks.
![Screenshot 2023-07-27 at 10 53 02 AM](https://github.com/aws/copilot-cli/assets/71282729/f82ce09f-19e9-4cf2-93cc-9d7930c86a5d)





By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.