chore: add option for only initialized workloads in selector (#5139)
This PR adds an options pattern to LocalWorkloadSelector so that it can grab all workloads defined locally, not just those represented in the config store. 

This is a prerequisite for initialization during deployment.

When you initialize a LocalWorkloadSelector without `selector.OnlyInitializedWorkloads`, you see something like this when you run `sel.Workload()`:
<img width="755" alt="Screen Shot 2023-07-28 at 12 39 44 PM" src="https://github.com/aws/copilot-cli/assets/25392995/40f9b763-6262-491f-ac73-3e0559e0c8c2">

This behavior will be reused for the `deploy` environment selector. 




By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.