feat: allow initializing workloads during copilot deploy (#5168)
This is stacked on top of a couple of PRs (like 5128 and 5139). The diff should get smaller over time as those are merged. 

Adds a new method to initialize which skips manifest writing. 

I think the ideal scenario here is actually to use an initializer option on the workload initializer struct, but I wanted to get this out quickly rather than doing another big 300-line refactor if possible. Happy to do this after release; i actually have started the work. 




By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.