chore: keep the pre/post actions' region as the pipeline region (#5220)
The CodeBuild project is deployed at the pipeline region. The action should look for the CodeBuild project at the same region (instead of the environment region, which can be different from the pipeline region).

With an upcoming change, people will still be able to access their environment VPC via. copilot commands from within the CodeBuild project, even if it lives in the pipeline region which is potentially different from the env region.


By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.