chore: remove empty buckets spinner (#5245)
Removes the spinner for emptying buckets. This is because most applications won't actually have any s3 buckets to delete, and showing this is highly misleading to users.


Follow up to #5184 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.