docs: clarify LBWS are public, BES are private (#5309)
Related: #5286, #5283.
While this does not _rename_ Load Balanced Web Services and Backend Services (a change with wider implications), it does add emphasis to the public/private difference between the two service types, both in the docs and in the interactive prompt hints.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.