Allow assume role chaining with federated users (source identity and session tags)
Depending on the original credentials used to invoke the copilot cli command,
the EnvManagerRole can only be assumed when it allows the original source identity
and transitive tags to be passed to the session.

These permissions should not be of harm when the user's session does not have
a source identity or transitive tags.