fix: pass env to svc cleaner (#5368)
Passes the current environment to the `opts.newSvcCleaner` function which fixes a bug where the incorrect `EnvManagerRole` is used to attempt to delete the static site S3 bucket.


Fixes #5367 

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.