chore: run local scheduler (#5369)
This PR is the start of a refactor to the `run local` that manages the running containers. The main goal of this refactor is to better support future features like watch and startup ordering.

The next refactor PR will integrate this scheduler with the `run local` `package cli` code.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.