chore: generalize watcher test (#5457)
The number of events received by the watcher appears to be system dependent, so this converts the expected and actual events received from an array to a set. This means that we don't expect an order of events, and we just need one copy of each event in order to succeed.

Was causing `make local-test` to fail, should be fixed.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.