fix: KMS access denied errors when downloading the static env file (#5514)
## Bug description
Since Copilot 1.32.0 cross account deployments are broken for all services (except static website) due to lacking permissions on the task execution role to perform the `kms:Decrypt` operation on the required key.

## Background
Since #5329 the artifacts are encrypted with a [Kms Key](https://github.com/bencehornak/copilot-cli/blob/1146bf995d687bfff615a5c7d35266fb8bcb0b56/internal/pkg/template/templates/app/cf.yml#L17-L63), which belongs to the App stack.

Due to the wrong configuration of the `ExecutionRole` resource if one deploys the workload to a different account than where the app stack is located, the download of the env file fails due to the lacking `kms:Decrypt` permission on the corresponding key.

https://github.com/bencehornak/copilot-cli/blob/67db7a2d4f94c69c836586f714341e587b1b8220/internal/pkg/template/templates/workloads/partials/cf/executionrole.yml#L39-L43

Note that this policy grants access to all keys in the _workload's_ account (which is not a great practice either, see #4628), which is wrong, considering that the key is located in the app's account.

## Fix
The policy was changed to the following:
```
- Effect: 'Allow'
  Action:
    - 'kms:Decrypt'
  Resource:
    - !Ref ArtifactKeyARN
```

where `ArtifactKeyARN` is the ARN of the Kms Key.

## TODO
- [x] There are many failing test cases, because the generated YAML changes. Is there a way to efficiently fix all of these test cases?
- [x] Would it make sense to add a test case, which prevents this bug in the feature? If yes, how? -> changed some test cases to test if the KMS key ARN gets passed to the stacks

## Conclusion

Fixes partially: #4628.

By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.