fix: revert changes to custom resources of passing hostedzoneids (#5540)
Sorry again for this mistake 🙇  I did not anticipate that it breaks the permissions required to create environment and services.

Previously we added the logic to pass all the HostedZoneId as params in the custom resource in #5315. 
But this introduces a bug as @Lou1415926 described in this issue https://github.com/aws/copilot-cli/issues/5535#issuecomment-1846327824

So reverting this changes as these are not necessary because `listHostedZonesByName` always fetches all the public hosted zones followed by private hosted zones. 

And also I think these modifications for filtering  of hosted zones are not required in custom resources because `listHostedZonesByName` api call always lists all the public hosted zones first and at last it lists private hosted zones. The logic we have in all the custom resources always works to fetch public hosted zone.
 
```  
const data = await appRoute53Client.listHostedZonesByName({
        DNSName: domainName,
        MaxItems: "1",
    }).promise();  
```