feat!: run local `--use-task-role` flag (#5529)
Implements the second method of retrieving the task role in the credentials chain: `ecs exec`

Breaking change:
The `--use-task-role` flag will default to `true`. This is a breaking change because in the event that neither method of receiving credentials works, `copilot run local` will fail by default. To mitigate this, we introduce a recommended action displayed when `--use-task-role` fails that gives the user 2 recommendations to proceed:
1. Specify `copilot run local --use-task-role=false` to disable the feature.
2. Use a provided yaml override to allow the credential retrieving process to work.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.