chore: fix watcher integ test (#5574)
OS differences caused integ test to succeed on linux and fail on mac. This generalizes the tests so we don't care about the order of rename operations and skip duplicate write operations. It also normalizes filepaths to the way they are handled in the `fsnotify` package using `filepath.Join` and `filepath.ToSlash`.



By submitting this pull request, I confirm that you can use, modify, copy, and redistribute this contribution, under the Apache 2.0 License.